pyproject.toml
[project]
name = "nav2py_template_controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.8.*"
dependencies = [
    "pyyaml",
]